+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Cull lights and occluders in the main world (#71).
+
 ## [0.10.0] - 2026-09-15
 
 ### Added